Validations
Validations on textbox data

+++ ProjectA/ProjectA/ManageAdvisor.cs
@@ -21,27 +21,38 @@ namespace ProjectA
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
+            
             SqlConnection conn = new SqlConnection(conURL);
             SqlCommand delcmd = new SqlCommand();
             int d = dataGridViewAD.CurrentCell.RowIndex;
             if (dataGridViewAD.Rows[d].Cells["Delete"].Selected)
             {
-                conn.Open();
-                int g = Convert.ToInt32(dataGridViewAD.Rows[d].Cells["Id"].Value);
-                delcmd.CommandText = "DELETE FROM Advisor WHERE id = '" + g + "' ";
-
-                delcmd.Connection = conn;
-                delcmd.ExecuteNonQuery();
-
-                using (SqlCommand dcl = new SqlCommand())
+                DialogResult result = MessageBox.Show("Do You Want to delete?", "Delete", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+                if (result.Equals(DialogResult.OK))
                 {
-                    dcl.CommandText = "DELETE FROM Person WHERE id=" + dataGridViewAD.Rows[d].Cells["id"].Value.ToString() + "";
-                    dcl.Connection = conn;
-                    dcl.ExecuteNonQuery();
+                    //Do something
+                    conn.Open();
+                    int g = Convert.ToInt32(dataGridViewAD.Rows[d].Cells["Id"].Value);
+                    delcmd.CommandText = "DELETE FROM Advisor WHERE id = '" + g + "' ";
+
+                    delcmd.Connection = conn;
+                    delcmd.ExecuteNonQuery();
+
+                    using (SqlCommand dcl = new SqlCommand())
+                    {
+                        dcl.CommandText = "DELETE FROM Person WHERE id=" + dataGridViewAD.Rows[d].Cells["id"].Value.ToString() + "";
+                        dcl.Connection = conn;
+                        dcl.ExecuteNonQuery();
+                    }
+                    dataGridViewAD.Rows.RemoveAt(dataGridViewAD.CurrentCell.RowIndex);
+                    MessageBox.Show("Row Deleted");
+                    conn.Close();
                 }
-                dataGridViewAD.Rows.RemoveAt(dataGridViewAD.CurrentCell.RowIndex);
-                MessageBox.Show("Row Deleted");
-                conn.Close();
+                else
+                {
+                    MessageBox.Show("Action canceled");
+                }
+                
             }
             if (dataGridViewAD.Rows[d].Cells["Edit"].Selected)
             {
@@ -54,42 +65,52 @@ namespace ProjectA
 
         private void button3_Click(object sender, EventArgs e)
         {
-            SqlConnection conn = new SqlConnection(conURL);
-            int d = dataGridViewAD.CurrentCell.RowIndex;
-            conn.Open();
-            int g = Convert.ToInt32(dataGridViewAD.Rows[d].Cells["Id"].Value);
-            string qry;
-            String gen = BoxGender.Text;
-            String des = BoxDesig.Text;
-            // gender
-            string gender = string.Format("SELECT id FROM Lookup Where Category ='GENDER' AND Value= '" + gen + "'");
-            SqlCommand cmd = new SqlCommand(gender, conn);
-            int gende = Convert.ToInt32(cmd.ExecuteScalar().ToString());
-            // Designation
-            string desig = string.Format("SELECT id FROM Lookup Where Category ='DESIGNATION' AND Value= '" + des + "'");
-            SqlCommand cmd5 = new SqlCommand(desig, conn);
-            int desi = Convert.ToInt32(cmd5.ExecuteScalar().ToString());
-            // insert 
-            qry = "UPDATE Person SET FirstName = '" + BoxFirstName.Text + "', LastName = '" + BoxLastName.Text + "', Contact = '" + BoxContactNo.Text + "', Email = '" + BoxEmail.Text + "', DateOfBirth = '" + dateTimePicker1.Value + "', Gender = '" + gende + "' WHERE id = '" + g + "' ";
-            SqlCommand cmd3 = new SqlCommand(qry, conn);
-            int i = cmd3.ExecuteNonQuery();
-
-            string PersonId = string.Format("Select id From Person Where Email = '{0}'", BoxEmail.Text);
-            SqlCommand cmd2 = new SqlCommand(PersonId, conn);
-            int id = Convert.ToInt32(cmd2.ExecuteScalar().ToString());
-            string query = "UPDATE Advisor SET Id = '" + id + "', Designation = '" +BoxDesig.Text + "', Salary = '"+BoxSalary.Text+"' WHERE Student.id = '" + g + "' ";
-            SqlCommand cmd4 = new SqlCommand(query, conn);
-            int k = cmd4.ExecuteNonQuery();
-            if (k > 0 && i > 0)
+            DialogResult result = MessageBox.Show("Do You Want to update?", "Update", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+            if (result.Equals(DialogResult.OK))
             {
-                MessageBox.Show("Update sucessfully");
+                //Do something
+                SqlConnection conn = new SqlConnection(conURL);
+                int d = dataGridViewAD.CurrentCell.RowIndex;
+                conn.Open();
+                int g = Convert.ToInt32(dataGridViewAD.Rows[d].Cells["Id"].Value);
+                string qry;
+                String gen = BoxGender.Text;
+                String des = BoxDesig.Text;
+                // gender
+                string gender = string.Format("SELECT id FROM Lookup Where Category ='GENDER' AND Value= '" + gen + "'");
+                SqlCommand cmd = new SqlCommand(gender, conn);
+                int gende = Convert.ToInt32(cmd.ExecuteScalar().ToString());
+                // Designation
+                string desig = string.Format("SELECT id FROM Lookup Where Category ='DESIGNATION' AND Value= '" + des + "'");
+                SqlCommand cmd5 = new SqlCommand(desig, conn);
+                int desi = Convert.ToInt32(cmd5.ExecuteScalar().ToString());
+                // insert 
+                qry = "UPDATE Person SET FirstName = '" + BoxFirstName.Text + "', LastName = '" + BoxLastName.Text + "', Contact = '" + BoxContactNo.Text + "', Email = '" + BoxEmail.Text + "', DateOfBirth = '" + dateTimePicker1.Value + "', Gender = '" + gende + "' WHERE id = '" + g + "' ";
+                SqlCommand cmd3 = new SqlCommand(qry, conn);
+                int i = cmd3.ExecuteNonQuery();
+
+                string PersonId = string.Format("Select id From Person Where Email = '{0}'", BoxEmail.Text);
+                SqlCommand cmd2 = new SqlCommand(PersonId, conn);
+                int id = Convert.ToInt32(cmd2.ExecuteScalar().ToString());
+                string query = "UPDATE Advisor SET Id = '" + id + "', Designation = '" + BoxDesig.Text + "', Salary = '" + BoxSalary.Text + "' WHERE Student.id = '" + g + "' ";
+                SqlCommand cmd4 = new SqlCommand(query, conn);
+                int k = cmd4.ExecuteNonQuery();
+                if (k > 0 && i > 0)
+                {
+                    MessageBox.Show("Update sucessfully");
+                }
+                else
+                {
+                    MessageBox.Show("Inserted Not Inserted");
+                }
+                PanelUpdateAdvisor.Hide();
+                conn.Close();
             }
             else
             {
-                MessageBox.Show("Inserted Not Inserted");
+                MessageBox.Show("Action canceled");
             }
-            PanelUpdateAdvisor.Hide();
-            conn.Close();
+            
         }
 
         private void ManageAdvisor_Load(object sender, EventArgs e)
