+++ ProjectA/ProjectA/ManageStudent.cs
@@ -21,27 +21,40 @@ namespace ProjectA
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
+            
             SqlConnection conn = new SqlConnection(conURL);
             SqlCommand delcmd = new SqlCommand();
             int d = dataGridViewSt.CurrentCell.RowIndex;
+            
             if (dataGridViewSt.Rows[d].Cells["Delete"].Selected)
             {
-                conn.Open();
-                int g = Convert.ToInt32(dataGridViewSt.Rows[d].Cells["Id"].Value);
-                delcmd.CommandText = "DELETE FROM Student WHERE id = '" + g + "' ";
-
-                delcmd.Connection = conn;
-                delcmd.ExecuteNonQuery();
-
-                using (SqlCommand dcl = new SqlCommand())
+                DialogResult result = MessageBox.Show("Do You Want to delete?", "Delete", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+                if (result.Equals(DialogResult.OK))
                 {
-                    dcl.CommandText = "DELETE FROM Person WHERE id=" + dataGridViewSt.Rows[d].Cells["id"].Value.ToString() + "";
-                    dcl.Connection = conn;
-                    dcl.ExecuteNonQuery();
+                    //Do something
+                    conn.Open();
+                    int g = Convert.ToInt32(dataGridViewSt.Rows[d].Cells["Id"].Value);
+                    delcmd.CommandText = "DELETE FROM Student WHERE id = '" + g + "' ";
+
+                    delcmd.Connection = conn;
+                    delcmd.ExecuteNonQuery();
+
+                    using (SqlCommand dcl = new SqlCommand())
+                    {
+                        dcl.CommandText = "DELETE FROM Person WHERE id=" + dataGridViewSt.Rows[d].Cells["id"].Value.ToString() + "";
+                        dcl.Connection = conn;
+                        dcl.ExecuteNonQuery();
+                    }
+                    dataGridViewSt.Rows.RemoveAt(dataGridViewSt.CurrentCell.RowIndex);
+                    MessageBox.Show("Row Deleted");
+                    conn.Close();
+
                 }
-                dataGridViewSt.Rows.RemoveAt(dataGridViewSt.CurrentCell.RowIndex);
-                MessageBox.Show("Row Deleted");
-                conn.Close();
+                else
+                {
+                    MessageBox.Show("Action Canceled");
+                }
+                
             }
             if (dataGridViewSt.Rows[d].Cells["Edit"].Selected)
             {
@@ -50,39 +63,51 @@ namespace ProjectA
                 PanelUpdateStudent.Show();
 
             }
+
         }
 
         private void button3_Click(object sender, EventArgs e)
         {
-            SqlConnection conn = new SqlConnection(conURL);
-            int d = dataGridViewSt.CurrentCell.RowIndex;
-            conn.Open();
-            int g = Convert.ToInt32(dataGridViewSt.Rows[d].Cells["Id"].Value);
-            string qry;
-            String gen = BoxGender.Text;
-            string gender = string.Format("SELECT id FROM Lookup Where Category ='GENDER' AND Value= '" + gen + "'");
-            SqlCommand cmd = new SqlCommand(gender, conn);
-            int gende = Convert.ToInt32(cmd.ExecuteScalar().ToString());
-            qry = "UPDATE Person SET FirstName = '" + BoxFirstName.Text + "', LastName = '" + BoxLastName.Text + "', Contact = '" + BoxContactNo.Text + "', Email = '" + BoxEmail.Text + "', DateOfBirth = '" + dateTimePicker1.Value + "', Gender = '" + gende + "' WHERE id = '" + g + "' ";
-            SqlCommand cmd3 = new SqlCommand(qry, conn);
-            int i = cmd3.ExecuteNonQuery();
-
-            string PersonId = string.Format("Select id From Person Where Email = '{0}'", BoxEmail.Text);
-            SqlCommand cmd2 = new SqlCommand(PersonId, conn);
-            int id = Convert.ToInt32(cmd2.ExecuteScalar().ToString());
-            string query = "UPDATE Student SET Id = '" + id + "', RegistrationNo = '" + BoxRegNo.Text + "' WHERE Student.id = '" + g + "' ";
-            SqlCommand cmd4 = new SqlCommand(query, conn);
-            int k = cmd4.ExecuteNonQuery();
-            if (k > 0 && i > 0)
+            DialogResult result = MessageBox.Show("Do You Want to Update?", "Update", MessageBoxButtons.OKCancel, MessageBoxIcon.Information);
+            if (result.Equals(DialogResult.OK))
             {
-                MessageBox.Show("Update sucessfully");
+                //Do something
+                SqlConnection conn = new SqlConnection(conURL);
+                int d = dataGridViewSt.CurrentCell.RowIndex;
+                conn.Open();
+                int g = Convert.ToInt32(dataGridViewSt.Rows[d].Cells["Id"].Value);
+                string qry;
+                String gen = BoxGender.Text;
+                string gender = string.Format("SELECT id FROM Lookup Where Category ='GENDER' AND Value= '" + gen + "'");
+                SqlCommand cmd = new SqlCommand(gender, conn);
+                int gende = Convert.ToInt32(cmd.ExecuteScalar().ToString());
+                qry = "UPDATE Person SET FirstName = '" + BoxFirstName.Text + "', LastName = '" + BoxLastName.Text + "', Contact = '" + BoxContactNo.Text + "', Email = '" + BoxEmail.Text + "', DateOfBirth = '" + dateTimePicker1.Value + "', Gender = '" + gende + "' WHERE id = '" + g + "' ";
+                SqlCommand cmd3 = new SqlCommand(qry, conn);
+                int i = cmd3.ExecuteNonQuery();
+
+                string PersonId = string.Format("Select id From Person Where Email = '{0}'", BoxEmail.Text);
+                SqlCommand cmd2 = new SqlCommand(PersonId, conn);
+                int id = Convert.ToInt32(cmd2.ExecuteScalar().ToString());
+                string query = "UPDATE Student SET Id = '" + id + "', RegistrationNo = '" + BoxRegNo.Text + "' WHERE Student.id = '" + g + "' ";
+                SqlCommand cmd4 = new SqlCommand(query, conn);
+                int k = cmd4.ExecuteNonQuery();
+                if (k > 0 && i > 0)
+                {
+                    MessageBox.Show("Update sucessfully");
+                }
+                else
+                {
+                    MessageBox.Show("Inserted Not Inserted");
+                }
+                PanelUpdateStudent.Hide();
+                conn.Close();
+
             }
             else
             {
-                MessageBox.Show("Inserted Not Inserted");
+                MessageBox.Show("Action Canceled");
             }
-            PanelUpdateStudent.Hide();
-            conn.Close();
+            
         }
 
         private void ManageStudent_Load(object sender, EventArgs e)
